feat: implement parent-child deduplication and enhance document formatting in search results
